backend/requirements-minimal.txt
# Minimal requirements for basic functionality
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-dotenv
python-multipart
httpx
